+++ classes/output/courseformat/state/section.php
@@ -36,19 +36,15 @@ class section extends \core_courseformat\output\local\state\section {
      * @return \stdClass data context for a mustache template
      */
     public function export_for_template(\renderer_base $output): \stdClass {
+        /** @var \stdClass $data */
         $data = parent::export_for_template($output);
         $data->parent = $this->section->parent;
         $data->parentid = $this->section->parent ? $this->format->get_modinfo()->get_section_info($this->section->parent)->id : 0;
 
-        if (!$this->format->get_course()->showsection0title && $this->section->section === 0) {
-            $data->collapsed = true;
-        } else {
-            $data->collapsed = (bool)$this->section->collapsed;
-        }
-
         // For sections that are displayed as a link do not print list of cms or controls.
         $showaslink = $this->section->collapsed == FORMAT_FLEXSECTIONS_COLLAPSED
             && $this->format->get_viewed_section() != $this->section->section;
+        $data->showaslink = $showaslink;
         $data->children = [];
         if ($this->section->section) {
             foreach ($this->format->get_modinfo()->get_section_info_all() as $s) {