+++ lang/en/format_flexsections.php
@@ -22,6 +22,8 @@
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
+$string['accordion'] = 'Accordion effect';
+$string['accordiondesc'] = 'When one section is expanded, collapse all others.';
 $string['addsections'] = 'Add section';
 $string['addsubsection'] = 'Add subsection';
 $string['backtocourse'] = 'Back to course \'{$a}\'';