+++ settings.php
@@ -49,4 +49,7 @@ if ($ADMIN->fulltree) {
     $settings->add(new admin_setting_configselect('format_flexsections/courseindexdisplay',
         get_string('courseindexdisplay', 'format_flexsections'),
         get_string('courseindexdisplaydesc', 'format_flexsections'), 0, $options));
+    $settings->add(new admin_setting_configcheckbox('format_flexsections/accordion',
+        get_string('accordion', 'format_flexsections'),
+        get_string('accordiondesc', 'format_flexsections'), 0));
 }